+++ apps/inference/models/statistic/raw_data_statistic.cpp
@@ -27,12 +27,12 @@ bool RawDataStatistic::parametersAreValid(Statistic* stat, int index) {
 
 void RawDataStatistic::setParameterAtPosition(double value, int row,
                                               int column) {
-  set(value, seriesAt(static_cast<uint8_t>(m_activePageIndex)),
-      relativeColumn(column), row, false, true);
+  set(value, seriesAt(m_activePageIndex), relativeColumn(column), row, false,
+      true);
 }
 
 double RawDataStatistic::parameterAtPosition(int row, int column) const {
-  int series = seriesAt(static_cast<uint8_t>(m_activePageIndex));
+  int series = seriesAt(m_activePageIndex);
   if (row >= numberOfPairsOfSeries(series)) {
     return NAN;
   }
@@ -40,8 +40,7 @@ double RawDataStatistic::parameterAtPosition(int row, int column) const {
 }
 
 void RawDataStatistic::deleteParametersInColumn(int column) {
-  clearColumn(seriesAt(static_cast<uint8_t>(m_activePageIndex)),
-              relativeColumn(column));
+  clearColumn(seriesAt(m_activePageIndex), relativeColumn(column));
 }
 
 bool RawDataStatistic::deleteParameterAtPosition(int row, int column) {
@@ -49,7 +48,7 @@ bool RawDataStatistic::deleteParameterAtPosition(int row, int column) {
     // Param is already deleted
     return false;
   }
-  int series = seriesAt(static_cast<uint8_t>(m_activePageIndex));
+  int series = seriesAt(m_activePageIndex);
   int numberOfPairs = numberOfPairsOfSeries(series);
   deletePairOfSeriesAtIndex(series, row);
   // DoublePairStore::updateSeries has handled the deletion of empty rows
